+++ test/next/jdbc/prepare_test.clj
@@ -9,7 +9,7 @@
   (:require [clojure.test :refer [deftest is testing use-fixtures]]
             [next.jdbc :as jdbc]
             [next.jdbc.test-fixtures
-             :refer [with-test-db ds jtds? postgres? sqlite?]]
+             :refer [with-test-db db ds jtds? mssql? postgres? sqlite?]]
             [next.jdbc.prepare :as prep]
             [next.jdbc.specs :as specs]))
 
@@ -74,7 +74,7 @@ INSERT INTO fruit (name, appearance) VALUES (?,?)
                  (conj result (count (jdbc/execute! t ["select * from fruit"]))))))))
     (is (= 4 (count (jdbc/execute! (ds) ["select * from fruit"])))))
   (testing "large batch insert"
-    (when-not (or (jtds?) (postgres?) (sqlite?))
+    (when-not (or (jtds?) (sqlite?))
       (is (= [1 1 1 1 1 1 1 1 1 13]
              (jdbc/with-transaction [t (ds) {:rollback-only true}]
                (with-open [ps (jdbc/prepare t ["
@@ -92,4 +92,30 @@ INSERT INTO fruit (name, appearance) VALUES (?,?)
                                                    {:batch-size 4
                                                     :large true})]
                    (conj result (count (jdbc/execute! t ["select * from fruit"]))))))))
+      (is (= 4 (count (jdbc/execute! (ds) ["select * from fruit"]))))))
+  (testing "return generated keys"
+    (when-not (mssql?)
+      (let [results
+            (jdbc/with-transaction [t (ds) {:rollback-only true}]
+              (with-open [ps (jdbc/prepare t ["
+INSERT INTO fruit (name, appearance) VALUES (?,?)
+"]
+                                           {:return-keys true})]
+                (let [result (prep/execute-batch! ps [["fruit1" "one"]
+                                                      ["fruit2" "two"]
+                                                      ["fruit3" "three"]
+                                                      ["fruit4" "four"]
+                                                      ["fruit5" "five"]
+                                                      ["fruit6" "six"]
+                                                      ["fruit7" "seven"]
+                                                      ["fruit8" "eight"]
+                                                      ["fruit9" "nine"]]
+                                                  {:batch-size 4
+                                                   :return-generated-keys true})]
+                  (conj result (count (jdbc/execute! t ["select * from fruit"]))))))]
+        (is (= 13 (last results)))
+        (is (every? map? (butlast results)))
+        ;; Derby and SQLite only return one generated key per batch so there
+        ;; are only three keys, plus the overall count here:
+        (is (< 3 (count results))))
       (is (= 4 (count (jdbc/execute! (ds) ["select * from fruit"])))))))